+++ tests/rest/PostIndexCept.php
@@ -3,4 +3,4 @@ $I = new RestTester($scenario);
 $I->wantToTest('Method Not Allowed response. ');
 $I->sendPOST('/');
 $I->seeResponseCodeIs(405);
-$I->seeHttpHeader('Allow', 'GET');
+$I->seeHttpHeader('Allow', 'OPTIONS, GET');